(maint) Move Puppet Agent helpers to spec_helper
 - This code should be available in the shared spec gem
